for #1238 Add ComQueryBackendHandlerFactory.
